Move properties out of Excel subfolder
