+++ tests/XReports.Tests/PropertyHandlers/Excel/ExcelDateTimeFormatPropertyExcelHandlerTest.cs
@@ -1,7 +1,7 @@
 using System;
 using FluentAssertions;
 using XReports.Models;
-using XReports.Properties.Excel;
+using XReports.Properties;
 using XReports.PropertyHandlers.Excel;
 using Xunit;
