+++ tests/XReports.Tests/PropertyHandlers/Html/DateTimeFormatPropertyHtmlHandlerTest.cs
@@ -4,7 +4,6 @@ using System.Threading;
 using FluentAssertions;
 using XReports.Models;
 using XReports.Properties;
-using XReports.Properties.Excel;
 using XReports.PropertyHandlers.Html;
 using Xunit;
